refactor: group types into separate files
