+++ src/types/objects.ts
@@ -1,0 +1,12 @@
+export type LockedObjectType = {
+  lockMovementX: boolean
+  lockMovementY: boolean
+  lockRotation: boolean
+  lockScalingFlip: boolean
+  lockScalingX: boolean
+  lockScalingY: boolean
+  lockSkewingX: boolean
+  lockSkewingY: boolean
+  hasControls: boolean
+  selectable: boolean
+}
